requirements.txt
Django==6.0.3
djangorestframework==3.17.1
celery==5.6.3
redis==7.4.0
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
django-redis==6.0.0
gunicorn==23.0.0
djangorestframework-simplejwt==5.3.1
onnxruntime==1.24.4
opencv-python==4.12.0.88
numpy==2.2.6
pillow==12.2.0
